requirements.txt
langchain==0.3.27
langgraph==0.6.6
llama-index  # see note below; stable version unspecified

fastapi==0.116.1
uvicorn==0.35.0
SQLAlchemy==2.0.43
sqlite-utils==3.35  # although npm-based; may not apply to pip
python-dotenv==1.1.1
loguru==0.7.3
APScheduler==3.11.0
typer==0.16.1
streamlit==1.48.1
